+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix a typo in `Synonym.type` setter leading to a potential bug when
   the given `type` is `None`.
 - Fix miscellaneous bugs found with `mypy`.
+- `fastobo` serializer crashing on namespace clauses because of a type
+  issue.
 
 ## [1.1.5] - 2020-01-25
 [1.1.5]: https://github.com/althonos/pronto/compare/v1.1.4...v1.1.5